requirements.txt
h5py==2.10.0
imgaug==0.4.0
ipython==7.34.0
keras==2.0.8
matplotlib==3.5.1
numpy
opencv_python==4.6.0.66
roboticstoolbox_python
scikit_image==0.19.3
scipy
setuptools==59.6.0
tensorflow==1.14.0
keras-cv==0.3.4
tensorflow-datasets==4.7.0
mrcnn==0.2
PyWavelets